requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.7.0
django-redis==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
djoser==2.0.3
environs==9.3.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==20.9
pep8==1.7.1
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.4.0
